Refactoring
