+++ lib/flexer/src/main/scala/org/enso/flexer/Macro.scala
@@ -25,7 +25,7 @@ object Macro {
     val superClassName = tree match {
       case Select(_, name) => name
       case _ =>
-        println("ERROR: Wrong shape 67")
+        println("ERROR: Wrong shape 68")
         println("Expected Select(_, name), got:")
         println(showRaw(tree))
         throw new Error("Wrong shape")
