+++ lib/flexer/src/main/scala/org/enso/flexer/automata/DFA.scala
@@ -3,7 +3,7 @@ package org.enso.flexer.automata
 import scala.collection.mutable
 
 case class DFA(
-  vocabulary: Vocabulary,
+  vocabulary: Dict,
   links: Array[Array[Int]],
   endStatePriorityMap: mutable.Map[Int, State.Desc]
 )
