+++ lib/flexer/src/main/scala/org/enso/flexer/automata/Pattern.scala
@@ -1,25 +1,19 @@
-package org.enso.flexer
+package org.enso.flexer.automata
 
 trait Pattern {
-  def |(that: Pattern) = Or(this, that)
-
-  def >>(that: Pattern) = Seq_(this, that)
-
-  def many: Pattern = Many(this)
-
-  def many1: Pattern = this >> many
+  def |(that: Pattern):  Pattern = Or(this, that)
+  def >>(that: Pattern): Pattern = Seq_(this, that)
+  def many:              Pattern = Many(this)
+  def many1:             Pattern = this >> many
 }
 
 case object None_ extends Pattern
 case object Pass  extends Pattern
 
-case class Ran(start: Int, end: Int) extends Pattern
-
-case class Or(left: Pattern, right: Pattern) extends Pattern
-
+case class Ran(start: Int, end: Int)             extends Pattern
+case class Or(left: Pattern, right: Pattern)     extends Pattern
 case class Seq_(first: Pattern, second: Pattern) extends Pattern
-
-case class Many(body: Pattern) extends Pattern
+case class Many(body: Pattern)                   extends Pattern
 
 object Pattern {
   def range(start: Char, end: Char): Pattern = Ran(start.toInt, end.toInt)
